refactor(core/presentation): Split forms interfaces into three files: - /forms/fields/interface.ts - /forms/inputs/interface.ts - /forms/layouts/interface.ts
