+++ app/scripts/modules/core/src/presentation/forms/inputs/index.ts
@@ -10,4 +10,5 @@ export * from './StringsAsOptions';
 export * from './TextAreaInput';
 export * from './TextInput';
 export * from './expression';
+export * from './interface';
 export * from './utils';
