--- app/scripts/modules/core/src/presentation/forms/interface.ts
@@ -1,61 +1,0 @@
-import { FieldProps } from 'formik';
-import * as React from 'react';
-
-import { IValidationCategory, IValidator } from './validation';
-
-/** These props are used by FormField and FormikFormField components */
-export interface IFieldLayoutPropsWithoutInput extends IValidationProps {
-  required?: boolean;
-  label?: React.ReactNode;
-  help?: React.ReactNode;
-  actions?: React.ReactNode;
-}
-
-/** These props are used by FieldLayout components, such as StandardFieldLayout */
-export interface IFieldLayoutProps extends IFieldLayoutPropsWithoutInput {
-  input: React.ReactNode;
-}
-
-/**
- * These props are used by controlled components, such as <input> or Input components like TextInput
- * Some of the typings reference the typings supplied by formik FieldProps
- */
-export interface IControlledInputProps {
-  value: FieldProps['field']['value'];
-  onChange(e: React.ChangeEvent<any>): void;
-  onBlur: FieldProps['field']['onBlur'];
-  name: FieldProps['field']['name'];
-}
-
-export type Omit<T, K> = Pick<T, Exclude<keyof T, K>>;
-export type OmitControlledInputPropsFrom<T> = Omit<T, keyof IControlledInputProps>;
-
-/** These props are used by Input components, such as TextInput */
-export interface IValidationProps {
-  touched?: boolean;
-  validationMessage?: React.ReactNode;
-  validationStatus?: IValidationCategory;
-  addValidator?: (validator: IValidator) => void;
-  removeValidator?: (validator: IValidator) => void;
-}
-
-/** These props are used by Input components, such as TextInput */
-export interface IFormInputProps extends Partial<IControlledInputProps> {
-  validation?: IValidationProps;
-  inputClassName?: string;
-}
-
-/** These props are used by FormFields such as FormikFormField and FormField */
-export interface ICommonFormFieldProps {
-  input: React.ComponentType<IFormInputProps>;
-  layout?: React.ComponentType<IFieldLayoutProps>;
-}
-
-export interface IFormFieldApi {
-  name(): string;
-  label(): string;
-  value(): any;
-  touched(): boolean;
-  validationMessage(): React.ReactNode;
-  validationStatus(): IValidationCategory;
-}
