+++ app/scripts/modules/core/src/presentation/forms/layouts/LayoutContext.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
-import { IFieldLayoutProps } from '../interface';
+import { ILayoutProps } from './interface';
 import { StandardFieldLayout } from './StandardFieldLayout';
 
-export const LayoutContext = React.createContext<React.ComponentType<IFieldLayoutProps>>(StandardFieldLayout);
+export const LayoutContext = React.createContext<React.ComponentType<ILayoutProps>>(StandardFieldLayout);
 export const { Provider: LayoutProvider, Consumer: LayoutConsumer } = LayoutContext;
