+++ app/scripts/modules/core/src/presentation/forms/layouts/index.ts
@@ -1,3 +1,4 @@
 export * from './StandardFieldLayout';
 export * from './LayoutContext';
 export * from './ResponsiveFieldLayout';
+export * from './interface';
